requirements.txt
openai>=1.40.0
anthropic>=0.34.0
mistralai>=1.2.2
pydantic>=2.8.2
python-dotenv>=1.0.1
gspread>=6.1.2
google-auth>=2.33.0
orjson>=3.10.7
regex>=2024.7.24